Re-organizes the loop and creates Wake Up function
Re-arranges the actions within the loop to make it more readable and compact.

Includes a WakeUpSDS011() function to start the loop.

Moves the definition of header to the setup function.
Print_to_Serial() is changed to Print_results() and includes the printing of the header lines for visualizing the results via Serial monitor