Calculates the average efficiently
Creates a new function ComputeAvg() to calculate a running average. It includes all the requested data points and avoids the use of arrays to store the values.

Fixes the tab spacing between Time and PM2.5

Moves the declaration of variables for storing raw and average data to the global scope

Add comments on variable declarations and code blocks in the main file

Re-organizes the Loop function